Modifications to cmake and harvesting tools required to get GAMBIT with HDF5 compiling on my laptop, with all physics bits -Ditch'ed.
CMake changes:

1. Can now ditch hdf5 entirely with -Ditch="hdf5"
2. "sz" (szip) library removed from HDF5_LIBRARIES list in optional.cmake,
(if given by find_package(hdf5) command) since we don't need it and I was
getting linker errors from it.

Harvester changes:

1. These were required because -Ditching gm2calc did not ditch the frontend
headers and types for gm2calc, which meant that Eigen libraries were still
required. Now if SpecBit and gm2calc are ditched, Eigen is also not needed.
2. Essentially I just changed the scripts such that in the Backends
directory, during the searches for header and source files, the directories
called X_version will not get searched if X is -Ditch'ed.

I will test this back on Prometheus with the rest of the modules not
ditched, to make sure there is no problem if e.g. gm2calc is ditched but
the modules which use it aren't.
I guess this could be a problem since the types might still needed
in GAMBIT even without gm2calc, in which case a new solution to this
issue might be needed.

	modified:   Backends/scripts/backend_harvester.py
	modified:   CMakeLists.txt
	modified:   Utils/scripts/harvesting_tools.py
	modified:   cmake/optional.cmake
	modified:   cmake/scripts/update_cmakelists.py
	modified:   yaml_files/spartan.yaml